requirements.txt
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
psutil==7.2.2
py4j==0.10.9.9
pyparsing==3.3.2
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyyaml==6.0.3
six==1.17.0
tzdata==2026.1
